requirements.txt
pandas
python-dotenv
boto3
selenium
webdriver-manager
coloredlogs